requirements.txt
APScheduler==3.9.1
requests==2.28.1
loguru==0.6.0
python-dotenv==0.21.1
quarter_lib==0.0.25
python-telegram-bot==20.0
